+++ apps/ced-portal-be/src/adapters/outbound/drizzle/schema/tables.ts
@@ -266,7 +266,11 @@ export const placeMaterializedView = pgMaterializedView(
     country: varchar({ length: 255 }),
     id: char({ length: 26 }).notNull(),
     name: varchar({ length: 512 }),
+    operatorFiscalCode: varchar("operator_fiscal_code", {
+      length: 32,
+    }).notNull(),
     operatorId: char("operator_id", { length: 26 }).notNull(),
+    operatorName: varchar("operator_name", { length: 512 }).notNull(),
     postalCode: varchar("postal_code", { length: 20 }),
     profileDisplayName: varchar("profile_display_name", { length: 512 }),
     profileId: char("profile_id", { length: 26 }),
@@ -294,7 +298,11 @@ export const opportunityMaterializedView = pgMaterializedView(
     language: localizedMetadataLanguageEnum(),
     name: varchar({ length: 512 }),
     nationalTerritory: boolean("national_territory"),
+    operatorFiscalCode: varchar("operator_fiscal_code", {
+      length: 32,
+    }).notNull(),
     operatorId: char("operator_id", { length: 26 }),
+    operatorName: varchar("operator_name", { length: 512 }).notNull(),
     placeId: char("place_id", { length: 26 }),
     profileDisplayName: varchar("profile_display_name", { length: 512 }),
   },