+++ infra/resources/ced/prod/data.tf
@@ -25,3 +25,43 @@ data "azurerm_private_dns_zone" "privatelink_blob_core_windows_net" {
   name                = "privatelink.blob.core.windows.net"
   resource_group_name = module.azure_core_values.network_resource_group_name
 }
+
+data "azurerm_key_vault_secret" "dkim_record_1" {
+  name         = "ced-one-mail-dkim-1"
+  key_vault_id = data.azurerm_key_vault.common.id
+}
+
+data "azurerm_key_vault_secret" "dkim_record_2" {
+  name         = "ced-one-mail-dkim-2"
+  key_vault_id = data.azurerm_key_vault.common.id
+}
+
+data "azurerm_key_vault_secret" "dkim_record_3" {
+  name         = "ced-one-mail-dkim-3"
+  key_vault_id = data.azurerm_key_vault.common.id
+}
+
+data "azurerm_key_vault_secret" "cmf_mx_1_record" {
+  name         = "ced-one-mail-cmf-mx-1"
+  key_vault_id = data.azurerm_key_vault.common.id
+}
+
+data "azurerm_key_vault_secret" "cmf_txt_1_record" {
+  name         = "ced-one-mail-cmf-txt-1"
+  key_vault_id = data.azurerm_key_vault.common.id
+}
+
+data "azurerm_key_vault_secret" "cmf_mx_2_record" {
+  name         = "ced-one-mail-cmf-mx-2"
+  key_vault_id = data.azurerm_key_vault.common.id
+}
+
+data "azurerm_key_vault_secret" "cmf_txt_2_record" {
+  name         = "ced-one-mail-cmf-txt-2"
+  key_vault_id = data.azurerm_key_vault.common.id
+}
+
+data "azurerm_key_vault_secret" "dmarc_record" {
+  name         = "ced-one-mail-dmarc"
+  key_vault_id = data.azurerm_key_vault.common.id
+}